LPS-187979 Remove App.tsx and Rename ListAPIApplications to ViewAPIApplications
